backend/requirements.txt
fastapi
uvicorn
pytrends
pydantic
python-dotenv
supabase
google-genai
